Separated new world and open world windows and menu items.
