+++ yam/About.xaml.cs
@@ -27,7 +27,7 @@ namespace Yam
 
         #region Assembly Attribute Accessors
 
-        public string AssemblyTitle
+        public static string AssemblyTitle
         {
             get
             {
@@ -65,7 +65,7 @@ namespace Yam
             }
         }
 
-        public string AssemblyProduct
+        public static string AssemblyProduct
         {
             get
             {
@@ -78,7 +78,7 @@ namespace Yam
             }
         }
 
-        public string AssemblyCopyright
+        public static string AssemblyCopyright
         {
             get
             {
