+++ yam/NewWorldWindow.xaml
@@ -5,8 +5,89 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:Yam"
         mc:Ignorable="d"
-        Title="NewWorldWindow" Height="450" Width="800">
-    <Grid>
+        Title="NewWorldWindow" Height="399.8" Width="467.4" Background="{DynamicResource {x:Static SystemColors.ControlBrushKey}}">
+    <Grid Margin="10">
+        <Grid>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="Auto"/>
+                <ColumnDefinition Width="*"/>
+            </Grid.ColumnDefinitions>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="Auto"/>
+            </Grid.RowDefinitions>
+            <TextBlock Margin="10,0,0,10" Text="Connect to a new world" Grid.ColumnSpan="2" />
+            <TextBlock Grid.Column="0" Grid.Row="1" Margin="10,10,10,10" TextWrapping="Wrap" Text="World Name"
+                    HorizontalAlignment="Left" Height="Auto" VerticalAlignment="Center" Width="Auto"/>
+            <TextBox Grid.Column="1" Grid.Row="1" x:Name="worldNameText" HorizontalAlignment="Left" 
+                    Height="Auto" Margin="10,10,10,10" TextWrapping="Wrap" VerticalAlignment="Center" Width="120"/>
+
+            <TextBlock Grid.Column="0" Grid.Row="2" Margin="10,10,10,10" Text="World Address"
+                        HorizontalAlignment="Left" Height="Auto" VerticalAlignment="Center" Width="Auto">
+            </TextBlock>
+            <TextBox Grid.Column="1" Grid.Row="2" x:Name="worldURLText" HorizontalAlignment="Left" 
+                    Height="Auto" Margin="10,10,10,10" TextWrapping="Wrap" VerticalAlignment="Center" Width="120"/>
+
+
+            <TextBlock Grid.Column="0"  Grid.Row="3" Text="World Port"  Margin="10,10,10,10" TextWrapping="Wrap" 
+                HorizontalAlignment="Left" Height="Auto" VerticalAlignment="Center" Width="Auto">
+            </TextBlock>
+
+            <TextBox x:Name="worldPortText" Grid.Column="1"  Grid.Row="3" HorizontalAlignment="Left" 
+                Height="Auto" Margin="10,10,10,10" TextWrapping="Wrap" VerticalAlignment="Center" Width="120"/>
+
+            <CheckBox x:Name="loginCheck" Grid.Row="4" Margin="10" Grid.ColumnSpan="2"
+                   Checked="LoginCheck_Checked"
+	               Unchecked="LoginCheck_Unchecked">
+                <Grid>
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="1*"/>
+                        <ColumnDefinition Width="2*"/>
+                    </Grid.ColumnDefinitions>
+                    <Grid.RowDefinitions>
+                        <RowDefinition Height="Auto"/>
+                        <RowDefinition Height="Auto"/>
+                        <RowDefinition Height="Auto"/>
+                        <RowDefinition Height="Auto"/>
+
+                    </Grid.RowDefinitions>
+
+                    <TextBlock Grid.Row="0" Margin="10,0,0,10" Text="Automatically log in" Grid.ColumnSpan="2" />
+
+                    <TextBlock x:Name="usernameBlock" Grid.Row="1" Grid.Column="0" Margin="10,10,10,10" TextWrapping="Wrap" Text="User Name"
+                    HorizontalAlignment="Left" Height="Auto" VerticalAlignment="Center" Width="Auto"/>
+                    <TextBox x:Name="usernameText" Grid.Row="1" Grid.Column="1"  HorizontalAlignment="Left" 
+                    Height="Auto" Margin="10,10,10,10" TextWrapping="Wrap" VerticalAlignment="Center" Width="120"/>
+
+                    <TextBlock x:Name="passwordBlock" Grid.Row="2"  Grid.Column="0" Margin="10,10,10,10" TextWrapping="Wrap" Text="Password"
+                    HorizontalAlignment="Left" Height="Auto" VerticalAlignment="Center" Width="Auto"/>
+                    <TextBox Grid.Row="2"  Grid.Column="1" x:Name="passwordText" HorizontalAlignment="Left" 
+                    Height="Auto" Margin="10,10,10,10" TextWrapping="Wrap" VerticalAlignment="Center" Width="120"/>
+                </Grid>
+            </CheckBox>
+            <CheckBox x:Name="passwordCheck" Grid.Row="5" Margin="10,10,10,10" Grid.ColumnSpan="2"
+                  Checked="SaveLoginCheck_Checked"
+	               Unchecked="SaveLoginCheck_Unchecked">
+                <TextBlock Margin="10,0,10,10" 
+                           Text="Save World (note: password is saved as plain text)" 
+                           TextWrapping="Wrap"/>
+            </CheckBox>
+            <Separator Grid.Row="6" Grid.ColumnSpan="2" ></Separator>
+
+            <Button x:Name="OkNewWorldButton" Content="OK"
+                Grid.Row="8" HorizontalAlignment="Right" Margin="10,10,100,10"  VerticalAlignment="Center" 
+                Width="75" Click="OkNewWorldButton_Click" IsDefault="True"/>
+            <Button x:Name="CancelNewWorldButton" Content="Cancel" 
+                Grid.Row="8" HorizontalAlignment="Right" Margin="100,10,10,10"  VerticalAlignment="Center" 
+                Width="75" Click="CancelNewWorldButton_Click" IsCancel="True"/>
+
+        </Grid>
         
     </Grid>
 </Window>
