+++ yam/OpenWorldWindow.xaml
@@ -1,0 +1,36 @@
+﻿<Window x:Class="Yam.OpenWorldWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" ResizeMode="NoResize"
+        Title="Open New World" Height="285.55" Background="{DynamicResource {x:Static SystemColors.ControlBrushKey}}" Width="303.2"  >
+    <Grid Margin="0,0,-0,0">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+        </Grid.RowDefinitions>
+   
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="Auto"/>
+        </Grid.ColumnDefinitions>
+            
+            <Grid Width="281" Grid.Row="0">
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                </Grid.RowDefinitions>
+                <TextBlock Margin="10,10,0,10" Text="Connect to a saved world"/>
+                <ListBox HorizontalAlignment="Center" Height="100" Grid.Row="1" x:Name="worldList"
+                     VerticalAlignment="Center" Width="100" 
+                     Margin="10,10,10,10" >
+                </ListBox>
+            </Grid>
+            <Separator Grid.Row="1" Margin="0,10,0,10"></Separator>
+
+            <Button x:Name="OkNewWorldButton" Content="OK"
+                Grid.Row="2" HorizontalAlignment="Right" Margin="10,10,100,10"  VerticalAlignment="Center" 
+                Width="75" Click="OkNewWorldButton_Click" IsDefault="True"/>
+            <Button x:Name="CancelNewWorldButton" Content="Cancel" 
+                Grid.Row="2" HorizontalAlignment="Right" Margin="100,10,10,10"  VerticalAlignment="Center" 
+                Width="75" Click="CancelNewWorldButton_Click" IsCancel="True"/>
+        </Grid>
+</Window>
