--- yam/newWorld.xaml
@@ -1,112 +1,0 @@
-﻿<Window x:Class="Yam.OpenWorldWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" ResizeMode="NoResize"
-        Title="Open New World" Height="547" Width="307" Background="{DynamicResource {x:Static SystemColors.ControlBrushKey}}"  >
-    <Grid Margin="0,0,0,0">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-        </Grid.RowDefinitions>
-        <RadioButton Grid.Row="0" Margin="10" x:Name="newWorldButton">
-            <Grid>
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="Auto"/>
-                    <ColumnDefinition Width="*"/>
-                </Grid.ColumnDefinitions>
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                </Grid.RowDefinitions>
-                <TextBlock Margin="10,0,0,10" Text="Connect to a new world" Grid.ColumnSpan="2" />
-                <TextBlock Grid.Column="0" Grid.Row="1" Margin="10,10,10,10" TextWrapping="Wrap" Text="World Name"
-                    HorizontalAlignment="Left" Height="Auto" VerticalAlignment="Center" Width="Auto"/>
-                <TextBox Grid.Column="1" Grid.Row="1" x:Name="worldNameText" HorizontalAlignment="Left" 
-                    Height="Auto" Margin="10,10,10,10" TextWrapping="Wrap" VerticalAlignment="Center" Width="120"/>
-
-                <TextBlock Grid.Column="0" Grid.Row="2" Margin="10,10,10,10" Text="World Address"
-                        HorizontalAlignment="Left" Height="Auto" VerticalAlignment="Center" Width="Auto">
-                </TextBlock>
-                <TextBox Grid.Column="1" Grid.Row="2" x:Name="worldURLText" HorizontalAlignment="Left" 
-                    Height="Auto" Margin="10,10,10,10" TextWrapping="Wrap" VerticalAlignment="Center" Width="120"/>
-
-
-                <TextBlock Grid.Column="0"  Grid.Row="3" Text="World Port"  Margin="10,10,10,10" TextWrapping="Wrap" 
-                HorizontalAlignment="Left" Height="Auto" VerticalAlignment="Center" Width="Auto">
-                </TextBlock>
-
-                <TextBox x:Name="worldPortText" Grid.Column="1"  Grid.Row="3" HorizontalAlignment="Left" 
-                Height="Auto" Margin="10,10,10,10" TextWrapping="Wrap" VerticalAlignment="Center" Width="120"/>
-
-                <CheckBox x:Name="loginCheck" Grid.Row="4" Margin="10" Grid.ColumnSpan="2"
-                   Checked="LoginCheck_Checked"
-	               Unchecked="LoginCheck_Unchecked">
-                    <Grid>
-                        <Grid.ColumnDefinitions>
-                            <ColumnDefinition Width="1*"/>
-                            <ColumnDefinition Width="2*"/>
-                        </Grid.ColumnDefinitions>
-                        <Grid.RowDefinitions>
-                            <RowDefinition Height="Auto"/>
-                            <RowDefinition Height="Auto"/>
-                            <RowDefinition Height="Auto"/>
-                            <RowDefinition Height="Auto"/>
-
-                        </Grid.RowDefinitions>
-
-                        <TextBlock Grid.Row="0" Margin="10,0,0,10" Text="Automatically log in" Grid.ColumnSpan="2" />
-
-                        <TextBlock x:Name="usernameBlock" Grid.Row="1" Grid.Column="0" Margin="10,10,10,10" TextWrapping="Wrap" Text="User Name"
-                    HorizontalAlignment="Left" Height="Auto" VerticalAlignment="Center" Width="Auto"/>
-                        <TextBox x:Name="usernameText" Grid.Row="1" Grid.Column="1"  HorizontalAlignment="Left" 
-                    Height="Auto" Margin="10,10,10,10" TextWrapping="Wrap" VerticalAlignment="Center" Width="120"/>
-
-                        <TextBlock x:Name="passwordBlock" Grid.Row="2"  Grid.Column="0" Margin="10,10,10,10" TextWrapping="Wrap" Text="Password"
-                    HorizontalAlignment="Left" Height="Auto" VerticalAlignment="Center" Width="Auto"/>
-                        <TextBox Grid.Row="2"  Grid.Column="1" x:Name="passwordText" HorizontalAlignment="Left" 
-                    Height="Auto" Margin="10,10,10,10" TextWrapping="Wrap" VerticalAlignment="Center" Width="120"/>
-                    </Grid>
-                </CheckBox>
-                <CheckBox x:Name="passwordCheck" Grid.Row="5" Margin="10,10,10,10" Grid.ColumnSpan="2"
-                  Checked="SaveLoginCheck_Checked"
-	               Unchecked="SaveLoginCheck_Unchecked">
-                    <TextBlock Margin="10,0,10,10" 
-                           Text="Save World (note: password is saved as plain text)" 
-                           TextWrapping="Wrap"/>
-                </CheckBox>
-            </Grid>
-
-        </RadioButton>
-        <Separator Grid.Row="2" HorizontalAlignment="Stretch" Height="1" Margin="0" VerticalAlignment="Top" Width="Auto"/>
-
-        <RadioButton x:Name="savedWorldButton" Grid.Row="3" Margin="10" HorizontalAlignment="Stretch" >
-            <Grid Width="281">
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                </Grid.RowDefinitions>
-                <TextBlock Margin="10,0,0,10" Text="Or connect to a saved world"/>
-                <ListBox HorizontalAlignment="Center" Height="50" Grid.Row="1" x:Name="worldList"
-                     VerticalAlignment="Center" Width="100" 
-                     Margin="10,10,10,10" >
-                </ListBox>
-            </Grid>
-        </RadioButton>
-        <Separator Grid.Row="4" ></Separator>
-
-        <Button x:Name="OkNewWorldButton" Content="OK"
-                Grid.Row="5" HorizontalAlignment="Right" Margin="10,10,100,10"  VerticalAlignment="Center" 
-                Width="75" Click="OkNewWorldButton_Click" IsDefault="True"/>
-        <Button x:Name="CancelNewWorldButton" Content="Cancel" 
-                Grid.Row="5" HorizontalAlignment="Right" Margin="100,10,10,10"  VerticalAlignment="Center" 
-                Width="75" Click="CancelNewWorldButton_Click" IsCancel="True"/>
-    </Grid>
-
-</Window>
